pyproject.toml
[project]
name = "moderaai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "brevo-python>=1.2.0",
    "cloudinary>=1.44.1",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "langchain-openai>=0.3.29",
    "pillow>=11.3.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
